@@ -34,6 +34,9 @@ sphinx = "^6.1.3"
 sphinxcontrib-log-cabinet = "^1.0.1"
 sphinx-rtd-theme = "^1.2.0"
 
+[tool.poetry.plugins."pytest11"]
+permaculture = "permaculture.testing.fixtures"
+
 [tool.poetry.plugins."permaculture.serializers"]
 "application/json" = "permaculture.serializer:json_serializer"
 "application/octet-stream" = "permaculture.serializer:octet_stream_serializer"
@@ -44,7 +47,8 @@ sphinx-rtd-theme = "^1.2.0"
 "text/plain" = "permaculture.serializer:text_plain_serializer"
 
 [tool.poetry.scripts]
-usda-plants = "permaculture.usda:main"
+usda-food = "permaculture.usda_food:main"
+usda-plants = "permaculture.usda_plants:main"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
@@ -52,11 +56,11 @@ build-backend = "poetry.core.masonry.api"
 
 [tool.black]
 line-length = 79
-target-version = ["py39"]
+target-version = ["py311"]
 preview = true
 
 [tool.ruff]
-target-version = "py39"
+target-version = "py311"
 line-length = 79
 fix = true
 select = [
